+++ client/src/hooks/useTextarea.js
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+
+const useTextarea = (initialValue) => {
+    const [value, setValue] = useState(initialValue);
+    const onChangeValue = (e) => {
+        setValue(e.target.value);
+    };
+
+    return [value, setValue, onChangeValue];
+};
+
+export default useTextarea;
